feat: Correção da lógica de agendamento. Criação da tela e a da lógica das solicitações do lado prestador.

+++ src/main/java/com/fixi/fixi/dto/request/RegistroPrestadorRequest.java
@@ -16,5 +16,5 @@ public class RegistroPrestadorRequest {
     private String tipoUsuario;
 
     // Lista de IDs ou nomes das categorias escolhidas
-    private List<Integer> categoriasIds;
+    private List<Long> categoriasIds;
 }
